requirements.txt
ollama>=0.3.0
pydantic>=2.0.0
rich>=13.0.0
playwright>=1.40.0
pyautogui>=0.9.54
python-telegram-bot>=21.0
python-dotenv>=1.0.0
